Move PopulatePropertiesTestHelper to core module
