+++ java/clients/src/test/java/sleeper/clients/admin/properties/PropertiesDiffTest.java
@@ -32,8 +32,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static sleeper.clients.admin.properties.PropertiesDiffTestHelper.newValue;
 import static sleeper.clients.admin.properties.PropertiesDiffTestHelper.valueChanged;
 import static sleeper.clients.admin.properties.PropertiesDiffTestHelper.valueDeleted;
-import static sleeper.clients.deploy.PopulatePropertiesTestHelper.generateTestInstanceProperties;
-import static sleeper.clients.deploy.PopulatePropertiesTestHelper.generateTestTableProperties;
+import static sleeper.core.deploy.PopulatePropertiesTestHelper.generateTestInstanceProperties;
+import static sleeper.core.deploy.PopulatePropertiesTestHelper.generateTestTableProperties;
 import static sleeper.core.properties.PropertiesUtils.loadProperties;
 import static sleeper.core.properties.instance.CommonProperty.LOG_RETENTION_IN_DAYS;
 import static sleeper.core.properties.instance.CommonProperty.MAXIMUM_CONNECTIONS_TO_S3;
